Fix Warnings

+++ src/data/corpus_sqlite.cpp
@@ -82,7 +82,7 @@ void CorpusSQLite::fillSQLite() {
       for(size_t i = 0; i < files_.size(); ++i) {
         cont = cont && utils::GetLine((std::istream&)*files_[i], line);
         if(cont)
-          ps.bind(i + 2, line);
+          ps.bind((int)(i + 2), line);
       }
 
       if(cont) {
@@ -114,7 +114,7 @@ SentenceTuple CorpusSQLite::next() {
     SentenceTuple tup(curId);
 
     for(size_t i = 0; i < files_.size(); ++i) {
-      auto line = select_->getColumn(i + 1);
+      auto line = select_->getColumn((int)(i + 1));
 
       if(i > 0 && i == alignFileIdx_) {
         addAlignmentToSentenceTuple(line, tup);
